Version010000 migration: Move non-schema changes to Version010000From000908 migration
- Add logging
- Fix migrateSystemTemplates() and migrateSystemThemes()
